+++ src/Addons/DrawnUi.Maui.Camera/Calculate.cs
