+++ src/Addons/DrawnUi.Maui.Camera/CameraEffect.cs
@@ -1,0 +1,12 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public enum CameraEffect
+{
+	None,
+	Grayscale,
+	ColorNegativeAuto,
+	ColorNegativeManual,
+	GrayscaleNegative,
+	Positive,
+	Debug
+}
