+++ src/Addons/DrawnUi.Maui.Camera/CameraPickerMode.cs
@@ -1,0 +1,10 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public enum CameraPickerMode
+{
+	None,
+	NegativeBlack,
+	NegativeGray,
+	NegativeWhite,
+	Error
+}
