+++ src/Addons/DrawnUi.Maui.Camera/CameraPosition.cs
@@ -1,0 +1,7 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public enum CameraPosition
+{
+	Default,
+	Selfie
+}
