+++ src/Addons/DrawnUi.Maui.Camera/CameraState.cs
@@ -1,0 +1,8 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public enum CameraState
+{
+	Off,
+	On,
+	Error
+}
