+++ src/Addons/DrawnUi.Maui.Camera/CameraType.cs
@@ -1,0 +1,7 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public enum CameraType
+{
+	Default,
+	Max
+}
