+++ src/Addons/DrawnUi.Maui.Camera/CaptureLocationType.cs
@@ -1,0 +1,12 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public enum CaptureLocationType
+{
+
+	Gallery,
+
+	Manual,
+
+	Bitmap
+
+}
