+++ src/Addons/DrawnUi.Maui.Camera/DeviceOrientation.cs
@@ -1,0 +1,12 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public enum DeviceOrientation
+{
+    Unknown,
+    Portrait,
+    PortraitUpsideDown,
+    LandscapeLeft,
+    LandscapeRight,
+    FaceUp,
+    FaceDown
+}
