+++ src/Addons/DrawnUi.Maui.Camera/ICameraViewRenderer.cs
@@ -1,0 +1,16 @@
+﻿using SkiaSharp;
+
+namespace DrawnUi.Maui.Camera;
+
+public interface ICameraViewRenderer
+{
+	void TakePicture();
+
+	Task<SKBitmap> TakePictureForSkia();
+
+	void FlashScreen();
+
+	//void AddFileToGallery(string filename);
+
+	Task<bool> SaveJpgStreamToGallery(Stream stream, string filename, double cameraSavedRotation, string exifFromAnotherFilename = null);
+}
