+++ src/Addons/DrawnUi.Maui.Camera/Metadata.cs
@@ -1,0 +1,13 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public class Metadata
+{
+    public int Orientation { get; set; }
+    public int ISO { get; set; }
+    public double FocalLength { get; set; }
+
+    public string Software { get; set; }
+    public string Vendor { get; set; }
+    public string Model { get; set; }
+
+}
