+++ src/Addons/DrawnUi.Maui.Camera/Platforms/MacCatalyst/SkiaCamera.Other.cs
@@ -1,0 +1,31 @@
+﻿
+
+namespace DrawnUi.Maui.Camera;
+
+public partial class SkiaCamera : SkiaControl
+{
+    public virtual void SetZoom(double value)
+    {
+
+    }
+    /// <summary>
+    /// Call on UI thread only
+    /// </summary>
+    /// <returns></returns>
+    public async Task<bool> RequestPermissions()
+    {
+        return true;
+    }
+
+    public void CommandToRenderer(string command)
+    {
+        throw new NotImplementedException();
+    }
+
+    public SKBitmap GetPreviewBitmap()
+    {
+        throw new NotImplementedException();
+    }
+
+
+}
