+++ src/Addons/DrawnUi.Maui.Camera/StretchModes.cs
@@ -1,0 +1,17 @@
+﻿namespace DrawnUi.Maui.Camera;
+
+public enum StretchModes
+{
+	/// <summary>
+	/// Whatever
+	/// </summary>
+	Default,
+	/// <summary>
+	/// Fit inside screen bounds
+	/// </summary>
+	Fit,
+	/// <summary>
+	/// Fullscreen
+	/// </summary>
+	Fill
+}
