+++ src/Addons/DrawnUi.Maui.Rive/Platforms/iOS/SkiaRive.Apple.cs
@@ -1,0 +1,13 @@
+﻿using DrawnUi.Maui.Draw;
+
+namespace DrawnUi.Maui.Rive;
+
+public partial class SkiaRive : SkiaControl
+{
+    public SkiaRive()
+    {
+        this.IsVisible = false;
+    }
+
+
+}
