+++ src/DrawnUi.Maui-dev.sln
@@ -33,6 +33,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DrawnUi.Maui.Game", "Addons
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DrawnUi.MauiGraphics", "Addons\DrawnUi.MauiGraphics\DrawnUi.MauiGraphics.csproj", "{FF6A6A51-01FC-4CEF-9386-D132AD9EBF18}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DrawnUi.Maui.Rive", "Addons\DrawnUi.Maui.Rive\DrawnUi.Maui.Rive.csproj", "{1F524822-400E-4801-AE05-4B3A14B603AD}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DrawnUi.Maui.Camera", "Addons\DrawnUi.Maui.Camera\DrawnUi.Maui.Camera.csproj", "{D4FC0BDB-3FE8-49C9-8B76-53C83128B899}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -75,6 +79,14 @@ Global
 		{FF6A6A51-01FC-4CEF-9386-D132AD9EBF18}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FF6A6A51-01FC-4CEF-9386-D132AD9EBF18}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FF6A6A51-01FC-4CEF-9386-D132AD9EBF18}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1F524822-400E-4801-AE05-4B3A14B603AD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1F524822-400E-4801-AE05-4B3A14B603AD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1F524822-400E-4801-AE05-4B3A14B603AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1F524822-400E-4801-AE05-4B3A14B603AD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D4FC0BDB-3FE8-49C9-8B76-53C83128B899}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D4FC0BDB-3FE8-49C9-8B76-53C83128B899}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D4FC0BDB-3FE8-49C9-8B76-53C83128B899}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D4FC0BDB-3FE8-49C9-8B76-53C83128B899}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -88,6 +100,8 @@ Global
 		{C8BAD2E9-FDFA-404D-85AE-2F179CA78016} = {507756AE-A224-4CDD-80C1-55F06502DE1B}
 		{C9CABC24-B5D7-4824-BD1B-279BEF419348} = {8B60E2B6-4347-450E-9999-319D8C7697D3}
 		{FF6A6A51-01FC-4CEF-9386-D132AD9EBF18} = {8B60E2B6-4347-450E-9999-319D8C7697D3}
+		{1F524822-400E-4801-AE05-4B3A14B603AD} = {8B60E2B6-4347-450E-9999-319D8C7697D3}
+		{D4FC0BDB-3FE8-49C9-8B76-53C83128B899} = {8B60E2B6-4347-450E-9999-319D8C7697D3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {08E673B7-06BF-468C-8D6C-2D1AA8E4A8CA}
