--- src/Engine/Controls/Camera/CameraEffect.cs
@@ -1,12 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public enum CameraEffect
-{
-	None,
-	Grayscale,
-	ColorNegativeAuto,
-	ColorNegativeManual,
-	GrayscaleNegative,
-	Positive,
-	Debug
-}
