--- src/Engine/Controls/Camera/CameraPickerMode.cs
@@ -1,10 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public enum CameraPickerMode
-{
-	None,
-	NegativeBlack,
-	NegativeGray,
-	NegativeWhite,
-	Error
-}
