--- src/Engine/Controls/Camera/CameraPosition.cs
@@ -1,7 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public enum CameraPosition
-{
-	Default,
-	Selfie
-}
