--- src/Engine/Controls/Camera/CameraState.cs
@@ -1,8 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public enum CameraState
-{
-	Off,
-	On,
-	Error
-}
