--- src/Engine/Controls/Camera/CameraType.cs
@@ -1,7 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public enum CameraType
-{
-	Default,
-	Max
-}
