--- src/Engine/Controls/Camera/CaptureLocationType.cs
@@ -1,12 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public enum CaptureLocationType
-{
-
-	Gallery,
-
-	Manual,
-
-	Bitmap
-
-}
