--- src/Engine/Controls/Camera/DeviceOrientation.cs
@@ -1,12 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public enum DeviceOrientation
-{
-	Unknown,
-	Portrait,
-	PortraitUpsideDown,
-	LandscapeLeft,
-	LandscapeRight,
-	FaceUp,
-	FaceDown
-}
