--- src/Engine/Controls/Camera/ICameraViewRenderer.cs
@@ -1,16 +1,0 @@
-﻿using SkiaSharp;
-
-namespace DrawnUi.Maui.Controls;
-
-public interface ICameraViewRenderer
-{
-	void TakePicture();
-
-	Task<SKBitmap> TakePictureForSkia();
-
-	void FlashScreen();
-
-	//void AddFileToGallery(string filename);
-
-	Task<bool> SaveJpgStreamToGallery(Stream stream, string filename, double cameraSavedRotation, string exifFromAnotherFilename = null);
-}
