--- src/Engine/Controls/Camera/Metadata.cs
@@ -1,13 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public class Metadata
-{
-	public int Orientation { get; set; }
-	public int ISO { get; set; }
-	public double FocalLength { get; set; }
-
-	public string Software { get; set; }
-	public string Vendor { get; set; }
-	public string Model { get; set; }
-
-}
