--- src/Engine/Controls/Camera/StretchModes.cs
@@ -1,17 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public enum StretchModes
-{
-	/// <summary>
-	/// Whatever
-	/// </summary>
-	Default,
-	/// <summary>
-	/// Fit inside screen bounds
-	/// </summary>
-	Fit,
-	/// <summary>
-	/// Fullscreen
-	/// </summary>
-	Fill
-}
