--- src/Engine/Platforms/Android/Controls/SkiaRive.Android.cs
@@ -1,11 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public partial class SkiaRive : SkiaControl
-{
-    public SkiaRive()
-    {
-        this.IsVisible = false;
-    }
-
-
-}
