--- src/Engine/Shared/SkiaRive.Apple.cs
@@ -1,11 +1,0 @@
-﻿namespace DrawnUi.Maui.Controls;
-
-public partial class SkiaRive : SkiaControl
-{
-    public SkiaRive()
-    {
-        this.IsVisible = false;
-    }
-
-
-}
